Refactor return type to increase accuracy of number

+++ StringCalculator.cpp
@@ -1,7 +1,7 @@
 #include "catch.hpp"
 #include "StringCalculator.hpp"
 
-int StringCalculator(std::string myString)
+double StringCalculator(std::string myString)
 {
 	return 0;
 }
